feat: add authorization verification to use cases

+++ src/application/use-cases/delete-contact.js
@@ -1,11 +1,20 @@
-class DeleteContactUseCase {
-  #contactsRepo;
+import { ForbiddenError } from '../../domain/errors/forbidden.js';
+import { Roles } from '../../domain/roles.js';
 
-  constructor({ contactsRepo }) {
+class DeleteContactUseCase {
+  #contactsRepo = null;
+  #sessionsService = null;
+
+  constructor({ contactsRepo, sessionsService }) {
     this.#contactsRepo = contactsRepo;
+    this.#sessionsService = sessionsService;
   }
 
   async exec(contactId) {
+    const user = await this.#sessionsService.getCurrentUser();
+    if (!user || user.role !== Roles.ADMIN) {
+      throw new ForbiddenError();
+    }
     return this.#contactsRepo.deleteById(contactId);
   }
 }
